+++ features/constants.barrel.ts
@@ -1,0 +1,24 @@
+// type
+
+export type { 
+    AuthStatus,
+    AuthTabKey,
+    OAuthProviderKey,
+    guestStatus,
+    UnifiedUser,
+    QueryLike,
+    SessionData,
+    AuthValidities,
+    AuthSessionUser,
+    AuthModalProps,
+} from "./auth/oauth/oauth.constants"
+
+
+// 상수
+
+export { 
+    AUTH_TABS,
+    OAUTH_PROVIDERS,
+    ACTIVE_PROVIDERS,
+    SUPPORTED_PROVIDERS,
+} from "./auth/oauth/oauth.constants"
